rename model.py to chat_template.py
