+++ lmdeploy/serve/turbomind/chatbot.py
@@ -15,7 +15,7 @@ import numpy as np
 import tritonclient.grpc as grpcclient
 from tritonclient.grpc.service_pb2 import ModelInferResponse
 
-from lmdeploy.model import MODELS
+from lmdeploy.chat_template import MODELS
 from lmdeploy.serve.turbomind.utils import (Postprocessor, Preprocessor,
                                             prepare_tensor)
 from lmdeploy.utils import filter_suffix
